Fixed implementation bugs. Fixed check for determining with listener to use. Original inferred merely inferred type. Moved global Twitch emote insert to base listener so it can be used for VODs. Some message fragments, like Twitch emotes, in VODs have additional layouts wrapping them (why Twitch??). Accounted for them in old frag removal.
